exported styles to separate file
