+++ src/components/Onboarding/ProgressBar/ProgressSection/index.js
@@ -1,21 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
-import { makeStyles } from "@material-ui/styles";
 
-const useStyles = makeStyles(theme => ({
-    number: {
-        borderRadius: 25,
-        padding: "3px 10px",
-        marginRight: 10,
-        backgroundColor: theme.palette.text.secondary,
-        color: theme.palette.text.white,
-    },
-    TabTitle: {
-        color: theme.palette.text.secondary,
-        fontSize: 14,
-        fontFamily: theme.typography.secondary.main,
-    },
-}));
+import { useStyles } from "./styles";
 
 export const ProgressStatusList = {
     IDLE: "idle",
