+++ src/components/Onboarding/index.js
@@ -8,28 +8,7 @@ import Authentication from "./Authentication";
 import TermsOfUse from "./TermsOfUse";
 import Session from "../../utility/stringConstants/Session";
 import WalletKey from "./WalletKey";
-
-const useStyles = theme => ({
-    onboardingContainer: {
-        paddingBottom: 40,
-        backgroundColor: theme.palette.text.gray8,
-    },
-    topSection: {
-        textAlign: "center",
-        "& h2": {
-            color: theme.palette.text.black1,
-            fontSize: 32,
-            fontWeight: theme.typography.fontweight,
-        },
-        "& p": {
-            margin: "20px 0 0",
-            color: theme.palette.text.gray3,
-            fontFamily: theme.typography.secondary.main,
-            fontSize: 20,
-            lineHeight: "30px",
-        },
-    },
-});
+import { useStyles } from "./styles";
 
 class Authorization extends Component {
     state = {
